Set up separate docker compose stack, rename ml backend services
