+++ docker-compose.yml
@@ -23,7 +23,6 @@ services:
       - postgres
       - redis
       - minio-init
-      - ml_backend
     volumes:
       - .:/app:z
     env_file:
@@ -34,6 +33,8 @@ services:
     ports:
       - "8000:8000"
     command: /start
+    networks:
+      - ml_network
 
   postgres:
     build:
@@ -142,14 +143,7 @@ services:
       - ./compose/local/minio/init.sh:/etc/minio/init.sh
     entrypoint: /etc/minio/init.sh
 
-  ml_backend:
-    build:
-      context: ./processing_services/custom
-    volumes:
-      - ./processing_services/custom/:/app
-    ports:
-      - "2005:2000"
-    networks:
-      default:
-        aliases:
-          - processing_service
+networks:
+  ml_network:
+    external: true
+    name: ml_network
