+++ processing_services/docker-compose.yml
@@ -1,20 +1,33 @@
 services:
-  example:
+  ml_backend_minimal:
+    build:
+      context: ./minimal
+    volumes:
+      - ./minimal/:/app:z
+    ports:
+      - "2000:2000"
+    extra_hosts:
+      - minio:host-gateway
+    networks:
+      ml_network:
+        aliases:
+          - ml_backend_minimal
+          - processing_service
+
+  ml_backend_example:
     build:
       context: ./example
     volumes:
       - ./example/:/app:z
     ports:
-      - "2000:2000"
+      - "2005:2000"
     extra_hosts:
       - minio:host-gateway
+    networks:
+      ml_network:
+        aliases:
+          - ml_backend_example
 
-  custom:
-    build:
-      context: ./custom
-    volumes:
-      - ./custom/:/app:z
-    ports:
-      - "2000:2000"
-    extra_hosts:
-      - minio:host-gateway
+networks:
+  ml_network:
+    name: ml_network
