+++ processing_services/example/api/algorithms.py
@@ -1,119 +1,159 @@
-from .schemas import AlgorithmCategoryMapResponse, AlgorithmConfigResponse
+import datetime
+import logging
+import random
 
-RANDOM_DETECTOR = AlgorithmConfigResponse(
-    name="Random Detector",
-    key="random-detector",
-    task_type="detection",
-    description="Return bounding boxes at random locations within the image bounds.",
-    version=1,
-    version_name="v1",
-    uri="https://huggingface.co/RolnickLab/random-detector",
-    category_map=None,
-)
+from .schemas import AlgorithmConfigResponse, AlgorithmReference, BoundingBox, ClassificationResponse, SourceImage
 
-CONSTANT_DETECTOR = AlgorithmConfigResponse(
-    name="Constant Detector",
-    key="constant-detector",
-    task_type="detection",
-    description="Return a fixed bounding box at a fixed location within the image bounds.",
-    version=1,
-    version_name="v1",
-    uri="https://huggingface.co/RolnickLab/constant-detector",
-    category_map=None,
-)
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.DEBUG)
 
-RANDOM_BINARY_CLASSIFIER = AlgorithmConfigResponse(
-    name="Random binary classifier",
-    key="random-binary-classifier",
-    task_type="classification",
-    description="Randomly return a classification of 'Moth' or 'Not a moth'",
-    version=1,
-    version_name="v1",
-    uri="https://huggingface.co/RolnickLab/random-binary-classifier",
-    category_map=AlgorithmCategoryMapResponse(
-        data=[
-            {
-                "index": 0,
-                "gbif_key": "1234",
-                "label": "Moth",
-                "source": "manual",
-                "taxon_rank": "SUPERFAMILY",
-            },
-            {
-                "index": 1,
-                "gbif_key": "4543",
-                "label": "Not a moth",
-                "source": "manual",
-                "taxon_rank": "ORDER",
-            },
-        ],
-        labels=["Moth", "Not a moth"],
-        version="v1",
-        description="A simple binary classifier",
-        uri="https://huggingface.co/RolnickLab/random-binary-classifier",
-    ),
-)
+SAVED_MODELS = {}
 
-CONSTANT_CLASSIFIER = AlgorithmConfigResponse(
-    name="Constant classifier",
-    key="constant-classifier",
-    task_type="classification",
-    description="Always return a classification of 'Moth'",
-    version=1,
-    version_name="v1",
-    uri="https://huggingface.co/RolnickLab/constant-classifier",
-    category_map=AlgorithmCategoryMapResponse(
-        data=[
-            {
-                "index": 0,
-                "gbif_key": "1234",
-                "label": "Moth",
-                "source": "manual",
-                "taxon_rank": "SUPERFAMILY",
-            }
-        ],
-        labels=["Moth"],
-        version="v1",
-        description="A classifier that always returns 'Moth'",
-        uri="https://huggingface.co/RolnickLab/constant-classifier",
-    ),
-)
 
-RANDOM_SPECIES_CLASSIFIER = AlgorithmConfigResponse(
-    name="Random species classifier",
-    key="random-species-classifier",
-    task_type="classification",
-    description="A random species classifier",
-    version=1,
-    version_name="v1",
-    uri="https://huggingface.co/RolnickLab/random-species-classifier",
-    category_map=AlgorithmCategoryMapResponse(
-        data=[
-            {
-                "index": 0,
-                "gbif_key": "1234",
-                "label": "Vanessa atalanta",
-                "source": "manual",
-                "taxon_rank": "SPECIES",
-            },
-            {
-                "index": 1,
-                "gbif_key": "4543",
-                "label": "Vanessa cardui",
-                "source": "manual",
-                "taxon_rank": "SPECIES",
-            },
-            {
-                "index": 2,
-                "gbif_key": "7890",
-                "label": "Vanessa itea",
-                "source": "manual",
-                "taxon_rank": "SPECIES",
-            },
-        ],
-        labels=["Vanessa atalanta", "Vanessa cardui", "Vanessa itea"],
-        version="v1",
-        description="A simple species classifier",
-        uri="https://huggigface.co/RolnickLab/random-species-classifier",
-    ),
-)
+class Algorithm:
+    algorithm_config_response: AlgorithmConfigResponse
+
+    def __init__(self):
+        if self.algorithm_config_response.key not in SAVED_MODELS:
+            logger.info(f"Compiling {self.algorithm_config_response.key}...")
+            self.compile()
+        else:
+            logger.info(f"Using existing model {self.algorithm_config_response.key}...")
+            self.model = SAVED_MODELS[self.algorithm_config_response.key]
+
+    def compile(self):
+        raise NotImplementedError("Subclasses must implement the compile method")
+
+    def run(self) -> list:
+        raise NotImplementedError("Subclasses must implement the run method")
+
+    algorithm_config_response = AlgorithmConfigResponse(
+        name="Base Algorithm",
+        key="base",
+        task_type="base",
+        description="A base class for all algorithms.",
+        version=1,
+        version_name="v1",
+        category_map=None,
+    )
+
+
+class LocalDetector(Algorithm):
+    """
+    A simple local detector that uses a constant bounding box for each image.
+    """
+
+    def compile(self):
+        pass
+
+    def run(self, source_image: SourceImage) -> list[BoundingBox]:
+        x1 = random.randint(0, source_image.width)
+        x2 = random.randint(0, source_image.width)
+        y1 = random.randint(0, source_image.height)
+        y2 = random.randint(0, source_image.height)
+
+        logger.info("Sending bounding box with coordinates {x1}, {y1}, {x2}, {y2}...")
+
+        return [
+            BoundingBox(
+                x1=min(x1, x2),
+                y1=min(y1, y2),
+                x2=max(x1, x2),
+                y2=max(y1, y2),
+            )
+        ]
+
+    algorithm_config_response = AlgorithmConfigResponse(
+        name="Local Detector",
+        key="local-detector",
+        task_type="detection",
+        description="A detector that uses a random bounding box for each image.",
+        version=1,
+        version_name="v1",
+        category_map=None,
+    )
+
+
+class ConstantDetector(Algorithm):
+    """
+    A simple local detector that uses a constant bounding box for each image.
+    """
+
+    def compile(self):
+        pass
+
+    def run(self, source_image: SourceImage) -> list[BoundingBox]:
+        x1 = source_image.width * 0.25
+        x2 = source_image.width * 0.75
+        y1 = source_image.height * 0.25
+        y2 = source_image.height * 0.75
+
+        logger.info(f"Sending bounding box with coordinates {x1}, {y1}, {x2}, {y2}...")
+
+        return [
+            BoundingBox(
+                x1=min(x1, x2),
+                y1=min(y1, y2),
+                x2=max(x1, x2),
+                y2=max(y1, y2),
+            )
+        ]
+
+    algorithm_config_response = AlgorithmConfigResponse(
+        name="Constant Detector",
+        key="constant-detector",
+        task_type="detection",
+        description="A detector that uses a constant bounding box for each image.",
+        version=1,
+        version_name="v1",
+        category_map=None,
+    )
+
+
+class LocalClassifier(Algorithm):
+    """
+    A simple local classifier that uses the Hugging Face pipeline to classify images.
+    """
+
+    def compile(self):
+        from transformers import pipeline
+
+        self.model = pipeline(model="google/vit-base-patch16-224")
+        SAVED_MODELS[self.algorithm_config_response.key] = self.model
+
+    def run(self, source_image: SourceImage) -> list[ClassificationResponse]:
+        # Define the algorithm compilation, execution
+        preds = self.model(images=source_image._pil)
+
+        labels = [pred["label"] for pred in preds]
+        scores = [pred["score"] for pred in preds]
+        max_score_index = scores.index(max(scores))
+        classification = labels[max_score_index]
+        logger.info(f"Classification: {classification}")
+        logger.info(f"labels: {labels}")
+        logger.info(f"scores: {scores}")
+        logger.info("Sending classification response...")
+
+        return [
+            ClassificationResponse(
+                classification=classification,
+                labels=labels,
+                scores=scores,
+                logits=scores,
+                timestamp=datetime.datetime.now(),
+                algorithm=AlgorithmReference(
+                    name=self.algorithm_config_response.name, key=self.algorithm_config_response.key
+                ),
+                terminal=True,
+            )
+        ]
+
+    algorithm_config_response = AlgorithmConfigResponse(
+        name="Local Classifier",
+        key="local-classifier",
+        task_type="classification",
+        description="A vision transformer model for image classification.",
+        version=1,
+        version_name="v1",
+        category_map=None,
+    )
