+++ processing_services/example/requirements.txt
@@ -3,3 +3,6 @@ uvicorn
 pydantic
 Pillow
 requests
+transformers
+torch
+torchvision
