+++ processing_services/minimal/Dockerfile
