+++ processing_services/minimal/main.py
