+++ processing_services/minimal/requirements.txt
@@ -3,6 +3,3 @@ uvicorn
 pydantic
 Pillow
 requests
-transformers
-torch
-torchvision
